esg/app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Databricks File Uploader - Upload files to Databricks volumes with ease"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.3.3",
    "streamlit>=1.50.0",
    "databricks-sdk>=0.19.0",
    "databricks-cli>=0.17.0",
    "python-dotenv>=1.0.0",
]
